Combine the standard stoich species view with the custom stoich species view.  Reafactored to support new reaction view.
