--- client/model-view/templates/editCustomStoichSpecie.pug
@@ -1,10 +1,0 @@
-div
-  div.inline
-    button.btn.btn-outline-secondary(class='btn-sm', data-hook='decrement') -
-  span.custom(data-hook="ratio")
-  div.inline
-    button.btn.btn-outline-secondary(class='btn-sm', data-hook='increment') +
-  select(data-hook="select-stoich-specie")
-    span.message.message-below.message-error(data-hook="message-container")
-      p(data-hook="message-text")
-  button.btn.btn-outline-secondary.custom(class='btn-sm', data-hook='remove') X
