--- client/model-view/templates/editStoichSpecie.pug
@@ -1,6 +1,0 @@
-div
-  label.select
-    span(data-hook="ratio")
-    select(data-hook="select-stoich-specie")
-      span.message.message-below.message-error(data-hook="message-container")
-        p(data-hook="message-text")
