--- client/model-view/views/edit-custom-stoich-specie.js
@@ -1,121 +1,0 @@
-/*
-StochSS is a platform for simulating biochemical systems
-Copyright (C) 2019-2022 StochSS developers.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-var $ = require('jquery');
-//views
-var SelectView = require('ampersand-select-view');
-//templates
-var template = require('../templates/editCustomStoichSpecie.pug');
-
-module.exports = SelectView.extend({
-  // SelectView expects a string template, so pre-render it
-  template: template(),
-  bindings: {
-    'model.ratio' : {
-      hook: 'ratio'
-    }
-  },
-  events: {
-    'change select' : 'selectChangeHandler',
-    'click [data-hook=increment]' : 'handleIncrement',
-    'click [data-hook=decrement]' : 'handleDecrement',
-    'click [data-hook=remove]' : 'deleteSpecie'
-  },
-  initialize: function (args) {
-    var self = this;
-    SelectView.prototype.initialize.apply(this, arguments);
-    this.value = this.model.specie || null;
-    this.isReactants = this.parent.parent.isReactants;
-    this.reactionType = this.parent.parent.reactionType;
-    this.stoichSpecies = this.parent.parent.collection;
-    this.stoichSpecies.on('add', function () {
-      self.toggleIncrementButton();
-    });
-    this.stoichSpecies.on('remove', function () {
-      self.toggleIncrementButton();
-    });
-  },
-  render: function () {
-    SelectView.prototype.render.apply(this);
-    this.toggleIncrementButton();
-    this.toggleDecrementButton();
-  },
-  selectChangeHandler: function (e) {
-    var species = this.getSpeciesCollection();
-    var reactions = this.getReactionsCollection();
-    var specie = species.filter(function (m) {
-      return m.name === e.target.selectedOptions.item(0).text;
-    })[0];
-    this.model.specie = specie;
-    this.value = specie;
-    reactions.trigger("change");
-    this.model.collection.parent.trigger('change-reaction')
-  },
-  getSpeciesCollection: function () {
-    return this.model.collection.parent.collection.parent.species;
-  },
-  getReactionsCollection: function () {
-    return this.model.collection.parent.collection;
-  },
-  handleIncrement: function () {
-    if(this.validateRatioIncrement()){
-      this.model.ratio++;
-      this.toggleIncrementButton();
-      this.parent.parent.toggleAddSpecieButton();
-      this.model.collection.parent.trigger('change-reaction')
-    }
-    this.toggleDecrementButton();
-  },
-  validateRatioIncrement: function () {
-    if(this.stoichSpecies.length < 2 && this.model.ratio < 2)
-      return true;
-    if(this.reactionType !== 'custom-massaction')
-      return true;
-    if(!this.isReactants)
-      return true;
-    return false;
-  },
-  toggleIncrementButton: function () {
-    if(!this.validateRatioIncrement()){
-      $(this.queryByHook('increment')).prop('disabled', true);
-    }else{
-      $(this.queryByHook('increment')).prop('disabled', false);
-    }
-  },
-  toggleDecrementButton: function () {
-    if(this.model.ratio <= 1)
-      $(this.queryByHook('decrement')).prop('disabled', true);
-    else
-      $(this.queryByHook('decrement')).prop('disabled', false);
-  },
-  handleDecrement: function () {
-    this.model.ratio--;
-    this.model.collection.parent.trigger('change-reaction')
-    this.toggleDecrementButton();
-    if(this.validateRatioIncrement()){
-      this.toggleIncrementButton();
-      this.parent.parent.toggleAddSpecieButton();
-    }
-  },
-  deleteSpecie: function () {
-    var reaction = this.model.collection.parent
-    this.collection.removeStoichSpecie(this.model);
-    reaction.trigger('change-reaction')
-    this.parent.parent.toggleAddSpecieButton();
-  },
-});
