--- client/model-view/views/edit-stoich-specie.js
@@ -1,62 +1,0 @@
-/*
-StochSS is a platform for simulating biochemical systems
-Copyright (C) 2019-2022 StochSS developers.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-var $ = require('jquery');
-//views
-var SelectView = require('ampersand-select-view');
-//templates
-var template = require('../templates/editStoichSpecie.pug');
-
-module.exports = SelectView.extend({
-  // SelectView expects a string template, so pre-render it
-  template: template(),
-  bindings: {
-    'model.ratio' : {
-      hook: 'ratio'
-    }
-  },
-  events: {
-    'change select' : 'selectChangeHandler'
-  },
-  initialize: function () {
-    SelectView.prototype.initialize.apply(this, arguments);
-    this.value = this.model.specie || null;
-  },
-  render: function() {
-    SelectView.prototype.render.apply(this, arguments);
-  },
-  update: function () {
-  },
-  selectChangeHandler: function (e) {
-    var species = this.getSpeciesCollection();
-    var reactions = this.getReactionsCollection();
-    var specie = species.filter(function (m) {
-      return m.name === e.target.selectedOptions.item(0).text;
-    })[0];
-    this.model.specie = specie;
-    this.value = specie;
-    reactions.trigger("change");
-    this.model.collection.parent.trigger('change-reaction')
-  },
-  getReactionsCollection: function () {
-    return this.model.collection.parent.collection;
-  },
-  getSpeciesCollection: function () {
-    return this.model.collection.parent.collection.parent.species;
-  },
-});
